+++ pybamm/models/submodels/active_material/loss_active_material.py
@@ -54,11 +54,12 @@ class LossActiveMaterial(BaseModel):
             )
         variables = self._get_standard_active_material_variables(eps_solid)
         lli_due_to_lam = pybamm.Variable(
-            f"Loss of lithium due to LAM in {domain} electrode [mol]"
+            "Loss of lithium due to loss of active material "
+            f"in {domain} electrode [mol]"
         )
         variables.update(
             {
-                "Loss of lithium due to LAM "
+                "Loss of lithium due to loss of active material "
                 f"in {domain} electrode [mol]": lli_due_to_lam
             }
         )
@@ -142,12 +143,13 @@ class LossActiveMaterial(BaseModel):
                 f"{Domain} electrode active material volume fraction change"
             ]
 
-        # Loss of lithium due to LAM
+        # Loss of lithium due to loss of active material
         # See eq 37 in "Sulzer, Valentin, et al. "Accelerated battery lifetime
         # simulations using adaptive inter-cycle extrapolation algorithm."
         # Journal of The Electrochemical Society 168.12 (2021): 120531.
         lli_due_to_lam = variables[
-            f"Loss of lithium due to LAM in {domain} electrode [mol]"
+            "Loss of lithium due to loss of active material "
+            f"in {domain} electrode [mol]"
         ]
         # Multiply by mol.m-3 * m3 to get mol
         c_s_av = variables[f"Average {domain} particle concentration [mol.m-3]"]
@@ -174,6 +176,7 @@ class LossActiveMaterial(BaseModel):
             self.initial_conditions = {eps_solid: eps_solid_init}
 
         lli_due_to_lam = variables[
-            f"Loss of lithium due to LAM in {domain} electrode [mol]"
+            "Loss of lithium due to loss of active material "
+            f"in {domain} electrode [mol]"
         ]
         self.initial_conditions[lli_due_to_lam] = pybamm.Scalar(0)