+++ tests/unit/test_simulation.py
@@ -149,34 +149,6 @@ class TestSimulation(unittest.TestCase):
         self.assertEqual(sim.parameter_values["Current function [A]"], 2 * current_1C)
         self.assertEqual(sim.C_rate, 2)
 
-    def test_set_external_variable(self):
-        model_options = {
-            "thermal": "lumped",
-            "external submodels": ["thermal", "negative primary particle"],
-        }
-        model = pybamm.lithium_ion.SPMe(model_options)
-        sim = pybamm.Simulation(model)
-
-        Nr = model.default_var_pts["r_n"]
-
-        T_av = 0
-        c_s_n_av = np.ones((Nr, 1)) * 0.5
-        external_variables = {
-            "Volume-averaged cell temperature": T_av,
-            "X-averaged negative particle concentration": c_s_n_av,
-        }
-
-        # Step
-        dt = 0.1
-        for _ in range(5):
-            sim.step(dt, external_variables=external_variables)
-        sim.plot(testing=True)
-
-        # Solve
-        t_eval = np.linspace(0, 3600)
-        sim.solve(t_eval, external_variables=external_variables)
-        sim.plot(testing=True)
-
     def test_step(self):
 
         dt = 0.001